// EmployeeManagement.Client/src/app/enums/toastr-position-type.ts
export enum ToastrPosition {
    TopRight = "toast-top-right",
    TopLeft = "toast-top-left",
    TopCenter = "toast-top-center",
    TopFullWidth = "toast-top-full-width",
    BottomRight = "toast-bottom-right",
    BottomLeft = "toast-bottom-left",
    BottomCenter = "toast-bottom-center",
    BottomFullWidth = "toast-bottom-full-width"
}